refactor: create separate package for routers (api/v1)
