+++ app/__init__.py
@@ -1,3 +1,7 @@
+from dotenv import load_dotenv
+
 from app.settings import Settings
+
+load_dotenv()
 
 settings = Settings()
